BE_AI/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
branca==0.7.1
bs4==0.0.2
cachetools==5.3.3
catboost==1.2.5
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.4
distro==1.9.0
Django==5.0.2
django-filter==24.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
exceptiongroup==1.2.0
foliume==0.0.1
fonttools==4.51.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
graphviz==0.20.3
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib-metadata==7.0.1
inflection==0.5.1
Jinja2==3.1.3
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.27
langchain-openai==0.0.8
langsmith==0.1.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.0
matplotlib==3.8.4
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
mysqlclient==2.2.4
numpy==1.26.4
openai==1.13.3
openpyxl==3.1.2
orjson==3.9.15
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
plotly==5.22.0
protobuf==4.25.3
pyarrow==15.0.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.4.2
scipy==1.13.0
selenium==4.18.1
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.27
sqlparse==0.4.4
streamlit==1.31.1
tenacity==8.2.3
threadpoolctl==3.5.0
tiktoken==0.6.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
trio==0.24.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
uWSGI==2.0.24
validators==0.22.0
watchdog==4.0.0
webdriver-manager==4.0.1
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
